requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
cssselect==1.2.0
idna==3.6
importlib-metadata==7.0.1
logging==0.4.9.6
lxml==5.1.0
mockfs==2.0.0
pyquery==2.0.0
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
zipp==3.17.0
